feat: add parallel FFN adapter to gpt_neox
